Split new and legacy syntax dev plugins
